backend/requirements.txt
django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
python-dotenv==1.0.1
ansible==9.2.0
ansible-core==2.16.3
apache-libcloud==3.4.1
argcomplete==3.1.4
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
Brlapi==0.8.5
certifi==2023.11.17
chardet==5.2.0
cliapp==1.20180812.1
click==8.1.6
cloud-init==25.2
cmdtest==0.32+git
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distlib==0.3.8
distro==1.9.0
distro-info==1.7+build1
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==2.1.4
fasteners==0.18
filelock==3.13.1
gyp==0.1
httplib2==0.20.4
idna==3.6
iotop==0.6
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.29.0
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
netaddr==0.8.0
ntlm-auth==1.5.0
oauthlib==3.2.2
olefile==0.46
packaging==24.0
paramiko==2.12.0
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
pipenv==2023.12.1
pipx==1.4.3
platformdirs==4.2.0
psutil==5.9.8
ptyprocess==0.7.0
pycairo==1.25.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pykerberos==1.1.14
PyNaCl==1.5.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
pytz==2024.1
pywinrm==0.4.3
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
requests-ntlm==1.1.0
resolvelib==1.0.1
rich==13.7.1
selinux @ file:///build/libselinux-IVxPh3/libselinux-3.5/src
setuptools==68.1.2
simplejson==3.19.2
six==1.16.0
ssh-import-id==5.11
systemd-python==235
texttable==1.6.7
ttystatus==0.38
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.16
userpath==1.9.1
virtualenv==20.25.0+ds
wadllib==1.3.6
websocket-client==1.7.0
wheel==0.42.0
xdg==5
xkit==0.0.0
xmltodict==0.13.0
zombie-imp==0.0.2
